Fixed a problem where TinyMCE editor was not editable after pasting into TinyMCE.
I tried to update TinyMCE to the latest version as well, but that
presented several additional problems so we'll delay that for now. NOTE:
This only glosses over a bug in the core TinyMCE editor. If we do
anything else with the paste plugin we'll want to just upgrade TinyMCE
completely.